Moved around check_date functions to the right mini-projects
